pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.48"
description = ""
authors = ["Albert Gevorkyan <agevork@yandex.ru>"]
packages = [
  { include = "gendiff" },
]


[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.0"
coverage = "^5.5"
pytest = "^6.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.script:main"
